Refactor vereinsmanschaftenübersich to manschaft
