+++ bogenliga/src/app/app.routing.ts
@@ -23,7 +23,7 @@ export const ROUTES: Routes = [
   {path: 'spotter', loadChildren: () => import('src/app/modules/spotter/spotter.module').then((m) => m.SpotterModule)},
   {path: 'schusszettel', loadChildren: () => import('./modules/schusszettel/schusszettel.module').then((m) => m.SchusszettelModule)},
   {path: 'hilfe', loadChildren: () => import('src/app/modules/hilfe/hilfe.module').then((m) => m.HilfeModule)},
-  {path: 'vereinsmannschaftenuebersicht', loadChildren: () => import('src/app/modules/vereinsmannschaftenuebersicht/vereinsmannschaftenuebersicht.module').then((m) => m.VereinsmannschaftenuebersichtModule)},
-  {path: 'vereinsmannschaftenuebersicht/:id', loadChildren: () => import('src/app/modules/vereinsmannschaftenuebersicht/vereinsmannschaftenuebersicht.module').then((m) => m.VereinsmannschaftenuebersichtModule)},
+  {path: 'mannschaft', loadChildren: () => import('src/app/modules/mannschaft/mannschaft.module').then((m) => m.MannschaftModule)},
+  {path: 'mannschaft/:id', loadChildren: () => import('src/app/modules/mannschaft/mannschaft.module').then((m) => m.MannschaftModule)},
   {path: 'wkdurchfuehrung/tabletadmin/:id/:id/:id', loadChildren: () => import('src/app/modules/spotter/spotter.module').then((m) => m.SpotterModule)},
 ];
