+++ bogenliga/src/app/modules/mannschaft/components/index.ts
@@ -1,0 +1,2 @@
+export * from './mannschaft/mannschaft.component';
+
