+++ bogenliga/src/app/modules/mannschaft/components/mannschaft/mannschaft.component.scss
