+++ bogenliga/src/app/modules/mannschaft/components/mannschaft/mannschaft.component.ts
@@ -7,7 +7,7 @@ import {DsbMannschaftDataProviderService} from "@verwaltung/services/dsb-mannsch
 import {DsbMannschaftDTO} from "@verwaltung/types/datatransfer/dsb-mannschaft-dto.class";
 import {VereinDataProviderService} from "@verwaltung/services/verein-data-provider.service";
 import {VereinDTO} from "@verwaltung/types/datatransfer/verein-dto.class";
-import {VEREINSMANNSCHAFTENUEBERSICHT_CONFIG, MANNSCHAFTEN_TABLE_CONFIG} from "./vereinsmannschaftenuebersicht.config";
+import {MANNSCHAFT_CONFIG, MANNSCHAFTEN_TABLE_CONFIG} from "./mannschaft.config";
 import {TableRow} from "@shared/components/tables/types/table-row.class";
 import {MannschaftTabelleDO} from "@verwaltung/types/mannschfttabelle-do.class";
 
@@ -18,12 +18,12 @@ const ID_PATH_PARAM = 'id';
 
 @Component({
   selector: 'bla-mannschaftsuebersicht',
-  templateUrl: './vereinsmannschaftenuebersicht.component.html',
-  styleUrls: ['./vereinsmannschaftenuebersicht.component.scss']
+  templateUrl: './mannschaft.component.html',
+  styleUrls: ['./mannschaft.component.scss']
 })
-export class VereinsmannschaftenuebersichtComponent extends CommonComponentDirective implements OnInit {
+export class MannschaftComponent extends CommonComponentDirective implements OnInit {
   private providedID: number | null = null;
-  public mannschaften: DsbMannschaftDTO[] | null = null;
+  public mannschaft: DsbMannschaftDTO | null = null;
   public verein: VereinDTO | null = null;
   public rows: TableRow[] | null = null;
   public loadingTable: boolean;
@@ -39,8 +39,6 @@ export class VereinsmannschaftenuebersichtComponent extends CommonComponentDirec
   }
 
   ngOnInit(): void {
-    // Intentionally left blank for now
-    console.log('Bin in Mannschaftsuebersicht');
     this.providedID = null;
 
     this.loading = true;
@@ -50,7 +48,6 @@ export class VereinsmannschaftenuebersichtComponent extends CommonComponentDirec
         this.providedID = parseInt(params[ID_PATH_PARAM], 10);
         console.log('This.providedID: ' + this.providedID);
         // Load after we have the ID
-        this.loadVereinsData();
         this.loadMannschaftData();
       } else {
         // no id provided
@@ -61,7 +58,7 @@ export class VereinsmannschaftenuebersichtComponent extends CommonComponentDirec
   loadVereinsData(): void{
     // set loading state
     this.loading = true;
-    this.vereinsDataProvider.findById(this.providedID)
+    this.vereinsDataProvider.findById(this.mannschaft.vereinId)
       .then((response: BogenligaResponse<VereinDTO>) => {
         console.log(response)
         this.verein = response.payload!;
@@ -76,12 +73,13 @@ export class VereinsmannschaftenuebersichtComponent extends CommonComponentDirec
   loadMannschaftData(): void{
     // set loading state
     this.loading = true;
-    this.mannschaftDataProvider.findAllByVereinsId(this.providedID)
-      .then((response: BogenligaResponse<DsbMannschaftDTO[]>) => {
+    this.mannschaftDataProvider.findById(this.providedID)
+      .then((response: BogenligaResponse<DsbMannschaftDTO>) => {
         console.log(response)
-        this.mannschaften = response.payload!;
+        this.mannschaft = response.payload!;
         this.loading = false;
-        this.mannschaftenrows();
+        //this.mannschaftenrows();
+        this.loadVereinsData();
       })
       .catch((response: BogenligaResponse<DsbMannschaftDTO>) => {
         console.error(response);
@@ -89,7 +87,7 @@ export class VereinsmannschaftenuebersichtComponent extends CommonComponentDirec
       });
   }
 
-  mannschaftenrows(): void{
+  /*mannschaftenrows(): void{
     this.mannschaften.forEach(mannschaft =>{
       const test: MannschaftTabelleDO = new MannschaftTabelleDO(mannschaft.name, mannschaft.id);
       this.tableContent.push(test)
@@ -98,7 +96,7 @@ export class VereinsmannschaftenuebersichtComponent extends CommonComponentDirec
     this.rows = toTableRows(this.tableContent);
     this.tableContent = [];
     this.loadingTable = false;
-  }
+  }*/
 
   public async getSelectedRow($event): Promise<void> {
     const rowValues = $event;
@@ -106,6 +104,6 @@ export class VereinsmannschaftenuebersichtComponent extends CommonComponentDirec
     this.router.navigate(['/mannschaftsuebersicht',rowValues.id]);
   }
 
-  readonly config = VEREINSMANNSCHAFTENUEBERSICHT_CONFIG;
+  readonly config = MANNSCHAFT_CONFIG;
   readonly config_table = MANNSCHAFTEN_TABLE_CONFIG;
 }
