+++ bogenliga/src/app/modules/mannschaft/components/mannschaft/mannschaft.config.ts
@@ -2,9 +2,9 @@ import {NavigationDialogConfig} from '@shared/components';
 import {TableConfig} from '@shared/components/tables/types/table-config.interface';
 import {TableActionType} from '@shared/components/tables/types/table-action-type.enum';
 
-export const VEREINSMANNSCHAFTENUEBERSICHT_CONFIG: NavigationDialogConfig = {
-  moduleTranslationKey: 'VEREINSMANNSCHAFTENUEBERSICHT',
-  pageTitleTranslationKey: 'VEREINSMANNSCHAFTENUEBERSICHT.TITLE',
+export const MANNSCHAFT_CONFIG: NavigationDialogConfig = {
+  moduleTranslationKey: 'MANNSCHAFT',
+  pageTitleTranslationKey: 'MANNSCHAFT.TITLE',
   navigationCardsConfig: {
     navigationCards: []
   },
@@ -14,7 +14,7 @@ export const MANNSCHAFTEN_TABLE_CONFIG: TableConfig = {
 
   columns: [
     {
-      translationKey: 'VEREINSMANNSCHAFTENUEBERSICHT.TABLE.HEADERS.MANNSCHAFTSNAME',
+      translationKey: 'MANNSCHAFT',
       propertyName:   'name',
       width:          20,
     },
