+++ bogenliga/src/app/modules/mannschaft/mannschaft.routing.ts
@@ -1,0 +1,7 @@
+import {Routes} from '@angular/router';
+import {MannschaftComponent} from "./components";
+
+export const MANNSCHAFTSUEBERSICHT_ROUTES: Routes = [
+  {path: '', pathMatch: 'full', component: MannschaftComponent}
+];
+
