--- bogenliga/src/app/modules/vereinsmannschaftenuebersicht/components/index.ts
@@ -1,2 +1,0 @@
-export * from './Vereinsmannschaftenuebersicht/vereinsmannschaftenuebersicht.component';
-
