--- bogenliga/src/app/modules/vereinsmannschaftenuebersicht/components/vereinsmannschaftenuebersicht/vereinsmannschaftenuebersicht.component.html
@@ -1,20 +1,0 @@
-<bla-navigation-dialog [config]="config">
-  <div class="vereinsmannschaftenuebersicht-page">
-
-    <!-- Loading indicator -->
-    <div *ngIf="loading">{{"VEREINSMANNSCHAFTENUEBERSICHT.STATUS.LOADING" | translate}}</div>
-
-    <!-- Mannschaft anzeigen -->
-    <div *ngIf="!loading && verein">
-      <h2>{{ verein.name || 'unbenannte Mannschaft' }} </h2>
-    </div>
-    <div class="row">
-      <bla-data-table
-        (onRowEntry)="getSelectedRow($event)"
-        [config]="config_table"
-        [loading]="loadingTable"
-        [rows]="rows">
-      </bla-data-table>
-    </div>
-  </div>
-</bla-navigation-dialog>
