--- bogenliga/src/app/modules/vereinsmannschaftenuebersicht/vereinsmannschaftenuebersicht.routing.ts
@@ -1,7 +1,0 @@
-import {Routes} from '@angular/router';
-import {VereinsmannschaftenuebersichtComponent} from "./components";
-
-export const MANNSCHAFTSUEBERSICHT_ROUTES: Routes = [
-  {path: '', pathMatch: 'full', component: VereinsmannschaftenuebersichtComponent}
-];
-
