+++ bogenliga/src/assets/i18n/en.json
@@ -2033,21 +2033,16 @@
                 }
               }
             },
-            "VEREINSMANNSCHAFTENUEBERSICHT": {
-              "TITLE": "Club overview",
-              "DESCRIPTION": "Here you will find a central overview of, all teams in th club.",
-              "PLACEHOLDER": "The club overview is currently under development. Data integration will take place shortly.",
+            "MANNSCHAFT": {
+              "TITLE": "Team overview",
+              "DESCRIPTION": "Here you will find a central overview a team.",
+              "PLACEHOLDER": "The team overview is currently under development. Data integration will take place shortly.",
               "STATUS": {
-                "LOADING": "Club overview is loading ...",
-                "EMPTY": "No club found. The data source is currently not providing any entries.",
-                "TIMEOUT": "The club overview could not be loaded (timeout). Please try again later.",
+                "LOADING": "Team overview is loading ...",
+                "EMPTY": "No team found. The data source is currently not providing any entries.",
+                "TIMEOUT": "The team overview could not be loaded (timeout). Please try again later.",
                 "OFFLINE_FALLBACK": "No live data available. Local reference data is displayed.",
-                "ERROR": "The club overview could not be loaded."
-              },
-              "TABLE": {
-                "HEADERS": {
-                  "MANNSCHAFTSNAME": "Team name"
-                }
+                "ERROR": "The team overview could not be loaded."
               }
             }
           }
